Move main to cmd folder
